[Core] Don't show versions without editURL (see #1046).
Description
-----------

With this PR I want to add the possibility to save versions in the `tl_version` table, which are not directly editable and therefore should not be  shown on the welcome page.

This makes the version table useable for extensions to save versions for extra data.

As far as I understand, the editUrl is always saved when a new version is created. So I simply added an extra WHERE statement. Another solution would be to have an extra `hide` column.

**To give an example for a use case:**
I have an extension which saves extra data related (and edited) with content elements in an extra table. When the content element is saved, a version is created and I use the `oncreate_version_callback` to save a version for the extra data, too. But these are not directly editable and makes the 'Last changes' overview unreadable.

Commits
-------

807bb0f1 Don't show versions without editURL