[Core] Do not always save a new version in the file/page tree widget (see #1184).
Description
-----------

The filetree and pagetree widgets force new versions every time the selection is empty. 

This is because the input of the order field is directly php exploded which returns always an array. An empty string will become an array with one empty entry (like `array(0=>'')`) and therefor is not equal to the order array extracted from the database in the `__construct` method which returns an empty result as an empty array.
This only happens if an `orderField` is specified in the dca.

This PR simply adds a check if the input is empty.

Commits
-------

a3b68142 Merge branch 'hotfix/4.4.7'
d9bb2c7f Check if the order input has a value
5b9bd181 Save empty order values for empty selections
6b7e6dc3 Fix the coding style.
81e57608 Fix the coding style.