[Core] Find models based on their type in the model argument resolver (see #1204)
Description
-----------

I create a controller for a content element and added `public function __invoke(ContentModel $model)`. Strangely, that lead to no output. After some digging, I found the dev debug bar output that told me the controller argument could not automatically be resolved. And therefore the fragment was not rendered. In the current implementation, the argument name **must** be `$contentModel` (because `contentModel` is the attribute name). That's pretty unfortunate, so I extended the argument resolver to find an argument by type. It works for two cases:

 1. `ContentElement $contentModel` searches request argument by that name
 2. Converts the class `ContentModel` to `contentModel` and searches request argument by that name. It's a simple `lcfirst` on the class name (without namespace).

Some explanations for the additional changes:
 - The scope attribute must have our underline prefix (`_scope`) to match the one we already have in the regular controllers. Initially we agreed not to name it like that because it means one would need a `$_scope` argument to the controller function. However, I came to realize that the developer actually should use the `ScopeMatcher` and never know the attribute name at all.
 - I changed the namespace because we already have a `HttpKernel` and `ArgumentResolverInterface` is a subnamespace of `HttpKernel` in the Symfony Framework. One less folder in the bundle …
 - Contrary to the previous implementation, the argument resolver only works if a request is in Contao scope. I think that's a reasonable change though, as we don't want to boot the framework otherwise.

Commits
-------

a724df05 Renamed scope attribute and improved model argument resolver
4c71a10e Remove the support for Model suffix in argument name
04d281e7 Renamed the service
af4a4301 Fix the coding style.