[Core] Do not make the response private when saving the session (see #1388).
Description
-----------

This PR fixes #1283.

Symfony has recently changed their session listener to always make the response private if the session has been started. Although we agree with the change, it renders the HTTP cache unusable, because Contao always starts a session (e.g. to store the user's language). This PR decorates the Symfony session listener and circumvents Symfony's changes by not making the response private if the request is a Contao front end request.

Of course, our long term goal should be to adjust Contao so it only starts a session if it really has to.

Commits
-------

6ecc5140 Set the front end response cache headers in an event listener (see #1283).
f937aafd Fix the copyright notice.
99ad48d8 Do not make the response private when saving the session (see #1388).
41580e22 Add the unit tests.
e9678569 Add a description to the session listener.