[Core] Allow multi-root and multi-page searches (see #1462).
Description
-----------

These changes allow you to expand/limit the search to multiple pages or website roots. This enables you to:

* limit the search to multiple pages and its children - instead of just one page and its children
* expand the search over multiple website roots - instead of being restricted to the current website root

The changes are fairly simple and straight forward. I tagged it as RFC since I am unsure whether the use of a new field called `searchPages` is really necessary. I decided to not use the existing `pages` field, since the sort order does not matter. Though may be I should simply alter the DCA in a `load_callback`, similar to `tl_content::setSingleSrcFlags`? Then the `pages` field could be reused for this.

Commits
-------

6feb7d49 allow multi-root search
8f11f965 use tl_module.pages
8d139932 Fix the coding style.