[Core] Prevent uploading large images (see #1507).
Description
-----------

Currently Contao uses the system settings `gdMaxImgWidth` and `gdMaxImgHeight` to prevent memory limit errors while processing large images. If an image is larger than any of those values, the image simply will not be processed and shown as is in the front end.

Despite instructing editors not to upload any large images it still happens that they do. This is problematic for the front end - because then you will have large images that a visitor has to download. And the image could break the layout, if the image is expected to be a certain size or aspect ratio.

With this PR you can disallow the back end upload of images larger than the maximum GD size setting.

![disallow](https://user-images.githubusercontent.com/4970961/39421931-8c3f7cf2-4c6b-11e8-8f7b-3fc2a1f5cff3.png)

This PR depends on fixing the issue of #1506.

Commits
-------

2fc73fdf disallow large images
668283ca use contao.image.reject_large_uploads parameter
118ffbfb Fix the coding style.