[Core] Add support for sub-palettes in personal profile (see #1559).
Description
-----------

The problem is that `Backend` executes `$this->objAjax->executePostActions($dc);` before the `callback` of the registered module is called. Actually not even before but it doesn't call the callback at all if there are ajax requests. So basically no ajax operation can work which is why I removed the callback which was not necessary anyway and put it into a `onload_callback` which is what I would do in a custom module anyway.

This means anything that's possible in a regular DC is now possible in the back end user profile as well.

Fixes https://github.com/contao/core-bundle/issues/1552

Commits
-------

52bfac51 Added support for subpalettes in personal profile
52cca135 Permission check and redirect
fd47c722 Moved the module access check to Backend
4531f7ea Improved extensibility of back end modules
486397a1 Fix the coding style.