[Manager] Use the manager plugin to get commands and features (see #75)
Description
-----------

As discussed with @leofeyer, the Contao Manager API should use *features* instead of a version number. This allows the Contao Manager to dynamically enable or disable UI options based on the supported features, which is way easier than testing for a number.

This PR allows Contao Manager Plugins to register API commands and announce features. This is more of a preparation because the Contao Manager obviously must know about a feature to actually use it, but at least this can be used by the `contao/manager-bundle` to enable application configuration (like `prepend_locale`).

/cc @contao/developers

Commits
-------

1abbaa6e Use manager plugin to get commands and features
7372b576 Remove the overrideApiFeatures method
1eaa853a Added features of manager-bundle and allow to get all env keys
0e379e14 Updated unit tests
5de72f99 Added unit tests for Application
3fc6c228 Added API feature to disable packages
d69a5436 Correctly set disabled packages in the API application
096703e5 Fix the coding style.