Only split the monorepo if the whole build is successful (see #2)
Description
-----------

The logic when the split runs on Travis CI is currently not very reasonable I think. If the tests fail in PHP 7.1 but are fine in PHP 7.2, the split still gets executed. Additionally the tests in PHP 7.2 run twice which is a waste of time and resources.

For such use cases travis has build stages that makes it possible to run jobs only if a group of other jobs succeeded.

~~I also changed the requirement for `contao/monorepo-tools` to a specific commit hash to avoid incompatibilities in future updates to the lib.~~

Commits
-------

671cb3d0 Run the monorepo split in a seperate stage
871c4f84 Don’t require specific version of monorepo-tools