Do not allow registering a Model with an empty primary key
The Registry should either categorically allow registering a model
without a primary key or categorically not allow it. This fix implements
the latter. It does not use `empty()` because that would overshoot and
prevent the primary key from being `0` or `"0"`, which may not be
recommendable, but does not have to be forbidden in this context.