Remove user agent css classes in fe_page by default (see #482)
Description
-----------

Personally, I have not a single Contao installation where I need these CSS classes.
If I really need to do things differently depending on the browser I always go for feature detection which is safer anyway.

The major disadvantage of using these classes is that they have to be different depending on the user agent that requests the page. This is why we've implemented rendering this insert tag as ESI fragment. This means that the whole page can be served from the cache and the reverse proxy then has to execute a separate request to the application only to get these CSS classes. On every single page!

I'm very convinced that for more than 90% of all the websites using Contao, these CSS classes are irrelevant which is why I think we should optimize the default behaviour for those.

People who really need these classes can still re-add `{{ua::class}}`. You don't even need to do it in the `fe_page` but you can add the insert tag to e.g. the layout if you know that it's only needed for a certain layout.

Commits
-------

150d68bf Remove user agent css classes in fe_page by default
ef24ca39 Update core-bundle/src/Resources/contao/templates/frontend/fe_page.html5
00e60b65 Remove the newline at the end of the file