Use the current firewall token if it matches the context (see #513)
Description
-----------

If the user is logged in in the current request (e.g. through *rememberme*), the session key does not exist, but the user exists in the firewall. We must always use/respect the current firewall if it matches the requested scope.

Commits
-------

d73f6890 Use the current firewall token if it matches the context
d0224347 Fixed unit tests
2d243f3d Use firewall names instead of user constants
7f49b835 CS fixer
862d94c9 Fixed unit tests
4d0b52be Added missing return argument
f4fb92ce Fix the unit tests and coding style