Remove a page's search index entries when its alias is changed (see #472)
Description
-----------

**How to reproduce the issue:**

* Create a page with the alias `foo`.

* Once you call `example.test/foo.html`, the page gets an entry in `tl_search`.

* Now change the alias to `bar`.

Currently, the entry with the now-outdated URL remains in `tl_search`, which means it shows up in the search results. It only gets removed once you call the outdated URL and get a 404, e. g. by visiting the outdated search result listing.

Also note that if you change only the alias, the page with the new URL does not get added once you visit it. If you change the alias and, e. g., the title, the new URL does get added, but the old one still remains.

**Suggested solution:**

A page's search index entries already get removed when the page is deleted[1]. We should do the same when the alias changes.

---
[1] https://github.com/contao/contao/commit/ba075d0d28c9c3f7de558b7b294fe1becf34e02d

Commits
-------

7274f39f Remove a page's search index entries when its alias is changed
507c26c6 Purge search index on alias change: Check for new page by VERSION_NUMBER
f4429a14 Remove page version check altogether