Remove the failure handler for 2FA authentication (see #676)
Description
-----------

The 2FA failure handler is incorrect in this case, because

1. It is only an implementation to log login errors, which should only happen to username/pw logins and not 2FA passwords
2. It extends the wrong parent class for 2FA authentication. The default 2FA failure handler uses a different firewall config flag for the redirect URL.

You can currently see the issue by enabling 2FA in the backend user, enter your username+password but not the 2FA code. Then go to `contao/two-factor`. It should redirect to `/contao/login` but it redirects to `/login` (the "random" default of Scheb/TwoFactorBundle`)

Commits
-------

d0a95f11 Remove the failure handler for 2FA authentication