Do not load the cache kernel in dev environment (see #712)
Description
-----------

This PR resolves https://github.com/contao/contao/issues/649. It won't load the Cache-Kernel while in `dev` environment. I'm not sure this is the way to go but it solves my problems with it.

Commits
-------

d1916ed7 Let the ContaoKernel handle whether cache should be used
67387c8b fixed tests
c10f8a53 added more tests to check the instance of the kernel
89ac0cee Adjust the unit test method names