Fix a potential problem with backend module permissions for regular users.
If one of the backend modules "in the middle" is unset, then the module permissions checkbox list will get incorrect values because the modules array will no longer be numeric. For example if we have the following system backend modules:

[system] => Array
(
    [0] => files
    [1] => log
    [2] => maintenance
    [3] => undo
    [4] => url_rewrites
)

The "undo" module is unset automatically for the backend module permissions field which results in an array:

[system] => Array
(
    [0] => files
    [1] => log
    [2] => maintenance
    [4] => url_rewrites
)

Thus the array is no longer numeric but associative and checkbox values in the field are being generated incorrectly (e.g. value="2" instead of value="maintenance").