Always set host and language when generating navigation (see #1458)
Description
-----------

Currently, the host and language is only forwarded to the `renderNavigation` method if it's different than the current one. But I don't really see a point in that, why would it not always be passed on? 

Also, the `$language` property isn't used at all anymore, it seems. Removing this condition is helpful to _get rid of_ `addLanguageToUrl` stuff.

Commits
-------

bdd68ade Always set host and language when generating navigation