Return the request argument if it's of the correct type (see #1640)
Description
-----------

We've added the `ModelArgumentResolver` for our fragments, where the `pageModel` request attribute is the ID of the origin page model.

Since CMF routing, the `pageModel` is also in the master request attribute, but there it's the actual page model instance. If the page is routed to a Symfony controller, injecting that attribute into the action (normally done by https://github.com/symfony/symfony/blob/master/src/Symfony/Component/HttpKernel/Controller/ArgumentResolver/RequestAttributeValueResolver.php) fails because of our resolver.

Commits
-------

bcaf4ce2 Return the request argument if it's of the correct type
11bd8509 Updated unit tests
565ddaa3 CS