Skip the initializeSystem hook if the temp folder does not yet exist (see #1685)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #929
| Docs PR or issue | -

<!--
Bugfixes should be based on the 4.4 or 4.9 branch and features on the master
branch. Select the correct branch in the "base:" drop-down menu above.

Replace this notice with a short README for your feature/bugfix. This will help
people to understand your PR and can be used as a start for the documentation.
-->

Commits
-------

e6a1c1f4 Skip the initializeSystem hook if the temp folder does not yet exist
f1de5eb4 Fix the logic