Fix playerAspect default value (see #1663)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | N/A
| Docs PR or issue | N/A

### Problem

When updating a Contao 3.5 installation, which has Contao's CSS framework enabled and contains YouTube videos, to 4.9, the YouTube videos will not be visible in the front end.

### Reproduction

1. Install Contao 3.5.40.
2. Enable all the stylesheets of the internal framework.
3. Create a YouTube video content element.
4. Update the installation to Contao 4.9.

The YouTube video will not be visible in the front end.

### Cause

The `sql` definition of `tl_content.playerAspect` is set to `varchar(8) NOT NULL default 'none'`, even though there is no `none` value in its `options` and it even has `includeBlankOption` enabled. When updating from Contao 3.5 or 4.4 for example, this field will be filled with the default value `none` and causes the following HTML in the front end:

```html
<div class="ce_youtube last block">
  <figure class="video_container">
    <div class="responsive ratio-none">
      <iframe src="https://www.youtube.com/embed/…" allowfullscreen="" width="640" height="360"></iframe>
    </div>
  </figure>
</div>
```

The `<div class="responsive ratio-none">` together with Contao's responsive CSS framework makes it invisible. When saving the content element once again, the correct empty string value will be set in the database.

Commits
-------

cba948ec fix wrong default value for tl_content.playerAspect
d52e8d70 Merge branch '4.9' into fix-youtube-player-aspect