Handle invalid language codes in the meta wizard (see #1694)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #1635
| Docs PR or issue | -

<!--
Bugfixes should be based on the 4.4 or 4.9 branch and features on the master
branch. Select the correct branch in the "base:" drop-down menu above.

Replace this notice with a short README for your feature/bugfix. This will help
people to understand your PR and can be used as a start for the documentation.
-->

Commits
-------

e4064ecd Handle invalid language codes in the meta wizard
1741f35f Merge branch '4.9' into fix/meta-wizard