Skip the temp folder check if there is no initialization hook (see #1690)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #1685
| Docs PR or issue | -

<!--
Bugfixes should be based on the 4.4 or 4.9 branch and features on the master
branch. Select the correct branch in the "base:" drop-down menu above.

Replace this notice with a short README for your feature/bugfix. This will help
people to understand your PR and can be used as a start for the documentation.
-->

Commits
-------

addf9f8f Skip the temp folder check if there is no initialization hook
448b85df Adjust the logic again