Set "information_schema_stats_expiry=0" in MySQL 8 (see #1700)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #958
| Docs PR or issue | -

<!--
Bugfixes should be based on the 4.4 or 4.9 branch and features on the master
branch. Select the correct branch in the "base:" drop-down menu above.

Replace this notice with a short README for your feature/bugfix. This will help
people to understand your PR and can be used as a start for the documentation.
-->

Commits
-------

6b398b20 Set "information_schema_stats_expiry=0" in MySQL 8