Add a "rawValue" property to the TextArea class (see #1714)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #736
| Docs PR or issue | -

<!--
Bugfixes should be based on the 4.4 or 4.9 branch and features on the master
branch. Select the correct branch in the "base:" drop-down menu above.

Replace this notice with a short README for your feature/bugfix. This will help
people to understand your PR and can be used as a start for the documentation.
-->

Commits
-------

5eca1512 Add a "rawValue" property to the TextArea class
69e58704 Also remove the str_replace() when returning the raw value