Re-add the redirect in the BackendUser::authenticate() method (see #1715)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | -
| Docs PR or issue | -

<!--
Bugfixes should be based on the 4.4 or 4.9 branch and features on the master
branch. Select the correct branch in the "base:" drop-down menu above.

Replace this notice with a short README for your feature/bugfix. This will help
people to understand your PR and can be used as a start for the documentation.
-->

Commits
-------

65d1c651 Re-add the redirect in the BackendUser::authenticate() method
9d485718 Run the CS fixer