Change deprecated-message to a more clearer sentence (see #1701)
Description
-----------

| Q                | A
| -----------------| ---
| Misleading deprecated message     | added a hint to the Symfony

After a quick talk on slack, the deprecated message for AbstractLockedCommand could be misleading (which "lock service"?). I suggest to change it to a more clearer message with the direct hint to the Symfony Lock component.

Commits
-------

c4d5610e Change deprecated-message for AbstractLockedCommand to a more clearer sentence.
8f2a71c5 Deprecated message for AbstractLockedCommand transferred to PHPDoc
01e9b0b8 Update core-bundle/src/Command/AbstractLockedCommand.php

typo corrected

Co-authored-by: Leo Feyer <github@contao.org>
9b7652b8 Merge branch '4.9' into 4.9