Always show the default template in the drop-down menu (see #1729)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #1060
| Docs PR or issue | -

Although this mostly works as discussed, there are edge cases like this:

<img width="628" alt="" src="https://user-images.githubusercontent.com/1192057/81574650-62d1eb00-93a6-11ea-9066-099ea620edca.png">

The screenshot shows that there is a custom `ce_text` template in the "Theme 2" theme. However, if the content element is used within "Theme 1", the default template will be used instead of the one of Theme 2, but the drop-down does not "say" so.

Not sure how to improve this, though.

Commits
-------

537cecdf Always show the default template in the drop-down menu