Correctly check for duplicate input parameters (see #1743)
Description
-----------

This fixes issue when manually matching requests, e.g. for the search indexer.

The mentioned issue (contao/core#4277) describes multiple cases, and it is where we added the _unused argument_ stuff. These lines were previously added to prevent duplicate parameters in the URL.
 1. For `http://example.com/foo/bar/foo/baz.html`, Contao (previously) correctly rendered the page with `foo` being set to `baz`. This also allowed for duplicate page URLs (e.g. `http://example.com/foo/bar/foo/bar/foo/bar/foo/baz.html` and the _unused argument_ handling wouldn't kick in.
 2. It should also not be possible to repeat a parameter that is already present in query parameters (e.g. `http://example.com/foo/bar.html?foo=baz`).

In Contao 3, this meant checking the `$_GET` array for duplicate data. In Symfony, this should only match on the current request.

**Warning:** what this PR cannot solve (and wasn't solved before) is actually running multiple requests that change the `$_GET` parameters 🤷

Commits
-------

9684fd22 Correctly check for duplicate input parameters
868950a4 CS