Fix addImageToTemplate with fullsize (see #1774)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | #1773 
| Docs PR or issue | N/A

Currently Contao passes an `stdClass` object to `addImageToTemplate` in three places: `ContentGallery`, `ContentVimeo` and `ContentYouTube`. Within `ContentGallery` it is never a problem, because the `fullsize` attribute will never be in `$arrItem`. However, in the other cases, the content element might contain `fullsize = '1'`, if the content element previously was of type `text` or `image` for example and then later on switched to `youtube` or `vimeo`.

I also added a check for `$arrItem['id']` since that might be empty too.

Commits
-------

df636277 check for Template instance (#1773)
6e181516 Merge branch '4.9' into fix-add-image-to-template-fullsize