Do not use the contao_root route (see #1869)
Description
-----------

The `contao_root` route simply points to the `/` path of the website. There's no need to generate such a route. There are only limited uses of it anyway in the core, and some actually cause issues. Removing the use of `contao_root` in the core will also aid in implementing #1516

1. This will actually fix the back end login link to the front end when in `preview.php`, because it will (correctly) remove the entry point script
2. Regarding the firewall configuration, the default value for `target` is `/` so this should have no effect (see https://github.com/symfony/symfony/blob/4.4/src/Symfony/Bundle/SecurityBundle/DependencyInjection/MainConfiguration.php#L214)

Commits
-------

e1ff1da6 Do not use the contao_root route
75a58211 Fixed unit test