Allow to override robots setting for news and events (see #1778)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #...
| Docs PR or issue | contao/docs#...

This pull request adds the ability to override robots setting in the news, events and faqs. If robots setting is overriden the robots tag is overriden and the item is excluded from xml sitemap if "noindex,nofollow" is chosen.

Known limitation: If a page is set to `noindex,nofollow` an item can't be included in the xml sitemap.

/cc @frontendschlampe

Commits
-------

2ed9d15e Allow to override robots setting

If robots setting is overriden the robots tag is overriden and the
item is excluded from xml sitemap if "noindex,nofollow" is chosen
21e7e3a7 Improve translation
87f82881 Fix a few minor issues
d57fb1cc Drop implementation for faqs

It will be added to #1779, see as requested in
https://github.com/contao/contao/pull/1778#pullrequestreview-437324426
8adcc5ee Drop leftover annotation