add exception for ContaoUserProviderTest testing with an invalid class
Here is why: The expected parameter is a UserInterface (enforced via UserProviderInterface) but only FrontendUser|BackendUser should be allowed. This type checking, however, is dynamically, so static analyzers cannot resolve it (= it must be annotated). The test is effectively validating the dynamic type check and therefore must be allowed to ignore the annotation.