Add CSV mime type and set Content-Type in File::sendToBrowser (see #2020)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | https://github.com/terminal42/contao-leads/issues/92
| Docs PR or issue | -

Currently, `File::sendToBrowser` does not set the `Content-Type` itself. Symfony's `ResponseListener` will execute `$response->prepare` which will try to guess the mime type, if the `Content-Type` header is missing in the response object. This uses PHP's _fileinfo_ component, which will guess the mime type based on the _contents_ of the file.

However, in case of CSV files for example, this will be `text/plain` instead of `text/csv`, the latter of which would be the standard according to [RFC 4180](https://tools.ietf.org/html/rfc4180).

So I think within the context of `\Contao\File::sendToBrowser`, the method should set the `Content-Type` according to Contao's own mime type mapping, if present, and otherwise let Symfony guess it as before.

This PR fixes this and adds the missing mime type for `.csv` files.

Commits
-------

f0bfb7d1 add CSV mime type and set content-type in File::sendToBrowser
e64a0e88 code style
47f5d6e4 always set the Content-Type