Do not add empty CSS classes to template (see #2115)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | -
| Docs PR or issue | -

Currently `$classes` will always be an array, passed down from `ModuleArticle::compile`. The array could contain `first` or `last` for example, but otherwise will be empty. In that case an empty string is added to the `class` template variable, resulting in HTML markup like this:

```html
<div class="ce_example_element ">
  …
</div>
```

This PR adds a check for whether the `$classes` array is empty, to prevent the superfluous space.

Commits
-------

e864b448 add empty check for classes
6abfbc95 only use empty