Add support for HTTP cache subscribers (see #2148)
Description
-----------

This adds support for https://github.com/contao/manager-plugin/pull/29 in manager-plugin 2.9.

According to @Toflar this should not be a problem if manager-plugin < 2.9 is installed. The interface name is just a symbol that does not exist, as long as no plugin implement it. If a plugin **does** implement it, it should require the plugin ^2.9.

Thanks to the new interface, it will be possible to add cache subscribers to do special handling that should apply to the http cache (as well). In my case, I need to implement geo localization for visitors and dynamically add _sort-of_ preflight header with the country, so responses can `Vary` on it.

I would love to have this in the LTS release, similar to https://github.com/contao/contao/pull/1982 this should not cause any conflicts with existing code or Contao functionality as it is totally separate.

Commits
-------

04e4489a Add support for HTTP cache subscribers